WIP: properly rename to JsonApiKit
